sft-pipeline/requirements.txt
asgiref==3.8.1
confluent-kafka==2.4.0
apache-airflow-providers-apache-kafka==1.5.0
pymilvus
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
boto3==1.38.29
torch>=2.0.0
sentence-transformers>=2.2.2
transformers>=4.30.0
peft>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
optuna>=3.0.0
dvc>=3.0.0
apache-airflow==2.9.2
apache-airflow-providers-amazon
apache-airflow-providers-postgres
apache-airflow-providers-ssh
